backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic[email]==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.31
psycopg[binary]==3.2.1
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
redis==5.0.7
rq==1.16.1
python-dateutil==2.9.0
ics==0.7.2
requests==2.32.3
email-validator==2.2.0
